Give access to CredClient in tests
There are a couple of tests that rely on creating a new project and/or
user within the test that rely on the identity_client. This doesn't work
when we want to test with api_v2=False.

As part of the isolated credentials we already have an abstraction layer
for these identity operations. Extract the abstraction class into its
own file and add it to clients.Manager as identity_utils (because I'm
not sure what better name to call it).

Change-Id: I6b3a505936f04d41df18c9599ca3c0b2ffd26aea