WIP: Use system creds for system operations
Keystone's is_admin_project and in future 'system' credentials are used
to perform operations that do not fall into the traditional project
hierarchy. These are operations such as hypervisor or infrastructure
commands that are independant of projects.

Currently to perform those operations you have to use the 'admin
project' and in future this will be renamed to a system scope.

Change-Id: Ic96ee6005606f4998d829ef651219bafdbaa95fe